+++ priv/repo/migrations/20200818073803_create_user_search_filters.exs
@@ -1,0 +1,15 @@
+defmodule TdDd.Repo.Migrations.CreateUserSearchFilters do
+  use Ecto.Migration
+
+  def change do
+    create table(:user_search_filters) do
+      add :name, :string
+      add :filters, :map
+      add :user_id, :integer
+
+      timestamps()
+    end
+
+    create unique_index(:user_search_filters, [:name, :user_id])
+  end
+end